Split the generator into a separate crate
