+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "crates/full-node/sov-stf-runner",
     "crates/full-node/sov-metrics",
     "crates/full-node/sov-eth-dev-signer",
+    "crates/full-node/sov-eth-dev-generator",
     "crates/full-node/sov-api-spec",
     "crates/full-node/full-node-configs",
     # Utils
@@ -139,6 +140,7 @@ sov-rollup-apis = { path = "crates/full-node/sov-rollup-apis", default-features 
 sov-api-spec = { path = "crates/full-node/sov-api-spec", default-features = false, version = "0.3" }
 full-node-configs = { path = "crates/full-node/full-node-configs", version = "0.3" }
 sov-eth-dev-signer = { path = "crates/full-node/sov-eth-dev-signer", default-features = false, version = "0.3" }
+sov-eth-dev-generator = { path = "crates/full-node/sov-eth-dev-generator", default-features = false, version = "0.3" }
 sov-modules-api = { path = "crates/module-system/sov-modules-api", default-features = false, version = "0.3" }
 sov-uniqueness = { path = "crates/module-system/module-implementations/sov-uniqueness", default-features = false, version = "0.3" }
 sov-paymaster = { path = "crates/module-system/module-implementations/sov-paymaster", default-features = false, version = "0.3" }
