+++ crates/full-node/sov-eth-dev-generator/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "sov-eth-dev-generator"
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+version = { workspace = true }
+readme = "README.md"
+publish = false
+
+[lints]
+workspace = true
+
+[dependencies]
+sov-eth-dev-signer = { workspace = true }
+revm = { workspace = true, features = ["std", "secp256k1"] }
+sov-transaction-generator = { workspace = true, optional = true }
+reth-primitives = { workspace = true, features = ["secp256k1"] }
+reth-rpc-types = { workspace = true }
+secp256k1 = { workspace = true }
+sha2 = { workspace = true, optional = true }
+rand_chacha = { workspace = true, optional = true }
+arbitrary = { workspace = true, optional = true }
+
+
+[features]
+default = []
+arbitrary = [
+    "dep:sov-transaction-generator",
+    "dep:sha2",
+    "dep:rand_chacha",
+    "dep:arbitrary",
+    "reth-primitives/arbitrary",
+    "reth-rpc-types/arbitrary",
+    "revm/arbitrary",
+]
