+++ crates/full-node/sov-eth-dev-generator/README.md
@@ -1,0 +1,3 @@
+# DevGenerator
+
+Simplifies generating random ethereum transactions in development.
