+++ crates/full-node/sov-eth-dev-generator/src/lib.rs
@@ -1,0 +1,2 @@
+mod randomness;
+pub mod transfer_generator;
