+++ crates/full-node/sov-eth-dev-signer/Cargo.toml
@@ -16,23 +16,6 @@ workspace = true
 [dependencies]
 thiserror = { workspace = true }
 revm = { workspace = true, features = ["std", "secp256k1"] }
-sov-transaction-generator = { workspace = true, optional = true }
 reth-primitives = { workspace = true, features = ["secp256k1"] }
 reth-rpc-types = { workspace = true }
 secp256k1 = { workspace = true }
-sha2 = { workspace = true, optional = true }
-rand_chacha = { workspace = true, optional = true }
-arbitrary = { workspace = true, optional = true }
-
-
-[features]
-default = []
-arbitrary = [
-    "dep:sov-transaction-generator",
-    "dep:sha2",
-    "dep:rand_chacha",
-    "dep:arbitrary",
-    "reth-primitives/arbitrary",
-    "reth-rpc-types/arbitrary",
-    "revm/arbitrary",
-]
