+++ deny.toml
@@ -206,6 +206,11 @@ expression = "LicenseRef-Sovereign-Permissionless-Commercial-License"
 license-files = []
 
 [[licenses.clarify]]
+name = "sov-eth-dev-generator"
+expression = "LicenseRef-Sovereign-Permissionless-Commercial-License"
+license-files = []
+
+[[licenses.clarify]]
 name = "sov-modules-api"
 expression = "LicenseRef-Sovereign-Permissionless-Commercial-License"
 license-files = []
